+++ src/DRIVE_AGAIN/drive.py
@@ -7,6 +7,7 @@ from dataclasses import dataclass
 import numpy as np
 
 from DRIVE_AGAIN.common import Command, Pose
+from DRIVE_AGAIN.dataset_reader import DatasetReader
 from DRIVE_AGAIN.dataset_recorder import DatasetRecorder
 from DRIVE_AGAIN.geofencing import Geofence
 from DRIVE_AGAIN.robot import Robot
@@ -151,6 +152,7 @@ class Drive:
 
         experience_dir = os.path.join("/home", "root", "datasets")
         self.dataset_recorder = DatasetRecorder(experience_dir)
+        self.dataset_reader = DatasetReader(experience_dir)
 
         self.commands = []
 
@@ -183,14 +185,14 @@ class Drive:
         self.dataset_recorder.save_experience(dataset_name)
 
     def get_datasets(self) -> list[str]:
-        return self.dataset_recorder.get_datasets()
+        return self.dataset_reader.get_datasets()
 
-    def load_geofence(self, dataset_name: str):
-        geofence_points = self.dataset_recorder.load_geofence(dataset_name)
+    def load_geofence(self, dataset_name: str, timestamp_ns: float):
+        geofence_points = self.dataset_reader.load_geofence(dataset_name)
         if geofence_points:
             geofence_points_tuple: list[tuple[float, float]] = [(point.x, point.y) for point in geofence_points]
             self.geofence = Geofence(geofence_points_tuple)
-            self.confirm_geofence(0)
+            self.confirm_geofence(timestamp_ns)
 
     def skip_current_step(self, timestamp_ns: float):
         logging.info("Skipping command...")
